+++ web/source/osk/oskViewComponent.ts
@@ -1,0 +1,7 @@
+namespace com.keyman.osk {
+  export interface OSKViewComponent {
+    readonly element: HTMLElement;
+    readonly layoutHeight: ParsedLengthStyle;
+    refreshLayout(): void;
+  }
+}